requirements.txt
mysql-connector-python>=8.0.0
numpy>=1.18.0
